% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_calc.R
\name{d_calc}
\alias{d_calc}
\title{Jost diversity calculation}
\usage{
d_calc(com.matrix, q, sample.weight)
}
\description{
Creates d_calc function that calculates Jost diversity for a dataframe
}
\examples{
d_calc()
}
\keyword{jost}
