% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_summarize.R
\name{d_summarize}
\alias{d_summarize}
\title{Jost diversity summary}
\usage{
d_summarize(com.matrix, q, sample.weight)
}
\description{
Uses Jost diversity calculations to get diversity values for alpha, beta, and gamma diversity of a dataset
}
\examples{
d_summarize()
}
\keyword{jost}
