% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jost_d.R
\name{jost_d}
\alias{jost_d}
\title{Jost diversity calculations}
\usage{
jost_d(q = 1, com.matrix, weighted = FALSE, boot = FALSE, n.sim = NULL)
}
\description{
Jost function to be called
}
\examples{
jost_d()
}
\keyword{jost}
