% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_rows.R
\name{normalize_rows}
\alias{normalize_rows}
\title{Normalize rows function}
\usage{
normalize_rows(com.matrix)
}
\description{
Normalizes rows for diversity calculations
}
\examples{
normalize_rows()
}
\keyword{jost}
